backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pdfplumber==0.11.4
groq==1.1.1
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
pytest==8.3.3
